20240104 @Mookse - message functions modularized - contribution functions modularized
